temporary disable due to jzy3d api changes
git-svn-id: https://svn.cebitec.uni-bielefeld.de/svn/mgx@1946 2f99df6c-41e9-6383-b5e2-de53e728a4e2

+++ MGX-Charts3D/src/main/java/de/cebitec/mgx/gui/charts3d/CustomLegendRenderer.java
@@ -33,8 +33,8 @@ class CustomLegendRenderer implements Renderer2d {
                 new AbstractTickProvider() {
 
                     @Override
-                    public float[] generateTicks(float min, float max, int steps) {
-                        return new float[]{0f, 1f, 2f};
+                    public double[] generateTicks(double min, double max, int steps) {
+                        return new double[]{0f, 1f, 2f};
                     }
 
                     @Override
@@ -45,7 +45,7 @@ class CustomLegendRenderer implements Renderer2d {
                 new ITickRenderer() {
 
                     @Override
-                    public String format(float value) {
+                    public String format(double value) {
                         switch ((int) value) {
                             case 0:
                                 return "Single-feat.";
